finalisation du projet avec de nouvelles fonctionnalitées ajoutées

+++ app/Http/Controllers/EvenementController.php
@@ -14,7 +14,7 @@ class EvenementController extends Controller
      */
     public function index()
     {
-        $evenements = Evenement::all();
+        $evenements = Evenement::orderBy('id','DESC')->get();
         return view('welcome', compact('evenements'));
     }
     /**
@@ -22,8 +22,12 @@ class EvenementController extends Controller
      */
     public function events()
     {
-        $evenements = Evenement::all();
-        return view('dashboard', compact('evenements'));
+        if (Auth::guard('association')->check()) {  
+            $evenements = Evenement::orderBy('id','DESC')->get();
+            return view('dashboard', compact('evenements'));
+        } else {
+            abort(403, "Vous ne pouvez pas accedez a cette page");
+        }
     }
 
     /**
@@ -44,7 +48,7 @@ class EvenementController extends Controller
             'libelle' => ['required', 'string', 'max:255'],
             'date_limite_inscription' => ['required', 'date'],
             'description' => ['required', 'string'],
-            'image_mise_en_avant' => ['required', 'mimes:jpeg,png,jpg,gif,svg|max:10000'],
+            'image_mise_en_avant' => ['required', 'mimes:jpeg,png,jpg,gif,svg'],
             'lieux' => 'required|string',
             'date_evenement' => ['required', 'date'],
         ]);
@@ -66,7 +70,7 @@ class EvenementController extends Controller
         if ($evenement->save()) {
             return redirect('/dashboard')->with("status", "Evenement ajouté avec succés");
         } else {
-            return back()->with("status", "Erreur lors de l'ajout de l'evenement veuillez reessayez svp");
+            return back()->with("error", "Erreur lors de l'ajout de l'evenement veuillez reessayez svp");
         }
     }
 
@@ -75,7 +79,7 @@ class EvenementController extends Controller
      */
     public function show()
     {
-        $evenements=Evenement::where('association_id',Auth::guard('association')->user()->id)->get();
+        $evenements=Evenement::where('association_id',Auth::guard('association')->user()->id)->orderBy('id','DESC')->get();
         return view('evenements.myevents',compact('evenements'));
         
     }
@@ -106,29 +110,34 @@ class EvenementController extends Controller
         ]);
 
         $evenement = Evenement::findOrFail($request->id);
-        if ($request->file('image_mise_en_avant')) {
-            if (File::exists(public_path('evenements/' . $evenement->image_mise_en_avant))) {
-                File::delete(public_path('images/' . $evenement->image_mise_en_avant));
+        // $this->authorize('update',$evenement,'association');
+        if ($evenement->association_id===Auth::guard('association')->user()->id) {
+            
+            if ($request->file('image_mise_en_avant')) {
+                if (File::exists(public_path('evenements/' . $evenement->image_mise_en_avant))) {
+                    File::delete(public_path('images/' . $evenement->image_mise_en_avant));
+                }
+                $file = $request->file('image_mise_en_avant');
+                $filename = date('YmdHi') . $file->getClientOriginalName();
+                $file->move(public_path('evenements'), $filename);
+                $evenement->image_mise_en_avant = $filename;
             }
-            $file = $request->file('image_mise_en_avant');
-            $filename = date('YmdHi') . $file->getClientOriginalName();
-            $file->move(public_path('evenements'), $filename);
-            $evenement->image_mise_en_avant = $filename;
+    
+    
+            $evenement->libelle = $request->libelle;
+            $evenement->date_limite_inscription = $request->date_limite_inscription;
+            $evenement->description = $request->description;
+            $evenement->lieux = $request->lieux;
+            $evenement->association_id = Auth::guard('association')->user()->id;
+            $evenement->date_evenement = $request->date_evenement;
+    
+            if ($evenement->update()) {
+                return redirect('/dashboard')->with("status", "Evenement modifié avec succés");
+            } else {
+                return back()->with("error", "Erreur lors de la modification de l'evenement veuillez reessayez svp");
+            }
         }
-
-
-        $evenement->libelle = $request->libelle;
-        $evenement->date_limite_inscription = $request->date_limite_inscription;
-        $evenement->description = $request->description;
-        $evenement->lieux = $request->lieux;
-        $evenement->association_id = Auth::guard('association')->user()->id;
-        $evenement->date_evenement = $request->date_evenement;
-
-        if ($evenement->update()) {
-            return redirect('/dashboard')->with("status", "Evenement modifié avec succés");
-        } else {
-            return back()->with("status", "Erreur lors de la modification de l'evenement veuillez reessayez svp");
-        }
+     abort(403,"Vous n'avez pas le droit pour modifier cet evenement");
     }
 
     /**
@@ -137,8 +146,14 @@ class EvenementController extends Controller
     public function cloture($id)
     {
         $evenement = Evenement::findOrFail($id);
+        if ($evenement->association_id===Auth::guard('association')->user()->id) {
+        // $this->authorize('update',$evenement,'association');
         $evenement->update(['est_cloture_ou_pas' => 1]);
-        return back()->with("status", "Evenement cloturé avec succés");;
+        return back()->with("status", "Evenement cloturé avec succés");
+        }
+        
+        abort(403,"Vous n'avez pas le droit pour modifier cet evenement");
+        
     }
 
     /**
@@ -147,7 +162,12 @@ class EvenementController extends Controller
     public function destroy($id)
     {
         $evenement = Evenement::findOrFail($id);
+        if ($evenement->association_id===Auth::guard('association')->user()->id) {
+        // $this->authorize('delete',$evenement,'association');
         $evenement->delete();
         return back()->with("status", "Evenement Supprimé avec succés");
+        }
+        
+     abort(403,"Vous n'avez pas le droit pour modifier cet evenement");
     }
 }
